+++ app/Http/Controllers/metodoEnvioController.php
@@ -6,6 +6,11 @@ use Illuminate\Http\Request;
 use Laracasts\Flash\Flash;
 use Illuminate\Support\Facades\Validator;
 use App\MedioEnvio;
+use App\Direccion;
+use App\Zona;
+use App\Parroquia;
+use App\Municipio;
+use App\Estado;
 
 class metodoEnvioController extends Controller
 {
@@ -14,11 +19,11 @@ class metodoEnvioController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index(Request $request)
+    public function index()
     {
-        $nombre=$request->get('nombre');
+       
     
-        $envio=MedioEnvio::where('nombre','like',"%$nombre%")->paginate(2);
+        $envio=MedioEnvio::All();
         return \view('plantilla.contenido.admin.metodoEnvio.consultar',compact('envio'));
     }
 
@@ -40,7 +45,6 @@ class metodoEnvioController extends Controller
      */
     public function store(Request $request)
     {
-    
      
         $v=Validator::make($request->all(),[
             'nombre'=>'required',
@@ -54,8 +58,13 @@ class metodoEnvioController extends Controller
         $envio=new MedioEnvio();
 
         $envio->nombre=$request->nombre;
-       
         $envio->tiempoEntrega=$request->tiempoEntrega;
+
+        if($request->alcance=='iribarren'){
+            $envio->dentroIribarren='si';
+        }else{
+            $envio->dentroIribarren='no';
+        }
 
         if($request->activo){
             $envio->status='A';
@@ -65,15 +74,29 @@ class metodoEnvioController extends Controller
 
         if($request->envioGratis){
             $envio->envioGratis='A';
-            $envio->precioEnvio=0;
+            $envio['0kgA30kg']=0;
+            $envio['31kgA50kg']=0;
+            $envio['50kgA100kg']=0;
+            $envio['101kgA200kg']=0;
+            $envio['mayorA201kg']=0;
         }else{
+
+            if($request->montoMinimo){
+                $envio['envioGratisApartir']=$request->montoMinimo;
+            }
+
             $envio->envioGratis='I';
-            $envio->precioEnvio=$request->precioEnvio;
+            $envio['0kgA30kg']=$request->precio0kg30kg;
+            $envio['31kgA50kg']=$request->precio31kg50kg;
+            $envio['50kgA100kg']=$request->precio51kg100kg;
+            $envio['101kgA200kg']=$request->precio101kg200kg;
+            $envio['mayorA201kg']=$request->precio201kg;
         }
+
        
         $envio->save();
 
-        \flash('Metodo de envio agregado con exito')->important()->warning();
+        \flash('Metodo de envio agregado con exito')->important()->success();
         return \redirect()->route('metodoEnvio.index');
     }
 
@@ -123,8 +146,13 @@ class metodoEnvioController extends Controller
         $envio=MedioEnvio::findOrFail($id);
 
         $envio->nombre=$request->nombre;
-        $envio->precioEnvio=$request->precioEnvio;
         $envio->tiempoEntrega=$request->tiempoEntrega;
+
+        if($request->alcance=='iribarren'){
+            $envio->dentroIribarren='si';
+        }else{
+            $envio->dentroIribarren='no';
+        }
 
         if($request->activo){
             $envio->status='A';
@@ -134,15 +162,27 @@ class metodoEnvioController extends Controller
 
         if($request->envioGratis){
             $envio->envioGratis='A';
-            $envio->precioEnvio=0;
+            $envio['0kgA30kg']=0;
+            $envio['31kgA50kg']=0;
+            $envio['50kgA100kg']=0;
+            $envio['101kgA200kg']=0;
+            $envio['mayorA201kg']=0;
         }else{
+
+            if($request->montoMinimo){
+                $envio['envioGratisApartir']=$request->montoMinimo;
+            }
+
             $envio->envioGratis='I';
-            $envio->precioEnvio=$request->precioEnvio;
+            $envio['0kgA30kg']=$request->precio0kg30kg;
+            $envio['31kgA50kg']=$request->precio31kg50kg;
+            $envio['50kgA100kg']=$request->precio51kg100kg;
+            $envio['101kgA200kg']=$request->precio101kg200kg;
+            $envio['mayorA201kg']=$request->precio201kg;
         }
-       
         $envio->save();
 
-        \flash('Metodo de envio actualizado con exito')->important()->success();
+        \flash('Método de envio actualizado con exito')->important()->success();
         return \redirect()->route('metodoEnvio.index');
     }
 
@@ -155,8 +195,27 @@ class metodoEnvioController extends Controller
     public function destroy($id)
     {
         $envio=MedioEnvio::findOrfail($id);
-        $envio->delete();
-        \flash('Metodo de envio eliminado con exito')->important()->success();
-        return \redirect()->route('metodoEnvio.index');
+        $pedidos=$envio->pedido;
+        if(count($pedidos)<=0){
+            return 'f';
+            $envio->delete();
+            \flash('Método de envio eliminado con exito')->important()->success();
+            return \redirect()->route('metodoEnvio.index');
+        }else{
+            
+            \flash('No puede eliminar este método de envio, a que posee pedidos asociados')->important()->warning();
+            return \redirect()->route('metodoEnvio.index');
+        }
+       
     }
+
+
+    public function rangoEnvio($id){
+        
+        $direccion=Direccion::find($id);
+        $municipio=$direccion->zona->parroquia->municipio;
+        return $municipio;
+
+    }
+
 }
